requirements.txt
Pillow
requests
selenium==3.14.0